Implemented Ghost site changed webhook handler
refs [MOM-300](https://linear.app/tryghost/issue/MOM-300/update-actor-data-with-ghost-site-settings#comment-4e49803f)

Implemented Ghost site changed webhook handler so that when the settings for a
Ghost instance are updated, the actor data is also updated in the database
(name, description, icon). For this to work, the Ghost instance needs to setup
a webhook for the `Site Changed` event and point it to the
`<host>/.ghost/activitypub/webhooks/site/changed` endpoint
