+++ package.json
@@ -35,6 +35,7 @@
     "@sentry/node": "8.13.0",
     "hono": "4.4.6",
     "knex": "3.1.0",
+    "ky": "1.4.0",
     "mysql2": "3.10.1",
     "uuid": "10.0.0",
     "x-forwarded-fetch": "0.2.0"
